Implemented selection outline scene:
The scene renders outlines using additional SubViewport and camera. It uses 2 shaders - 1 on a mesh covering the view and 1 on the SubViewportContainer.
- The first shader creates a map on the red color channel image with edges highlighted and non edges set to 0.
- The second shader uses that map to render the outline on the screen or to change its color to fully transparent.

Outlines are visible through the wall.

The outline system allows setting view masks to mark which objects should be outlined. Warning - Setting it to the mask visible by the main camera will cover the screen with the image from the first shader.