Fixed mouse_gesture sometimes getting stuck in rotating state:
The cancelation of the state required the release keybinding to be an exact match. This could cause the release not to be triggered when the user pressed the release button while holding some other modifier button (like shift). The cancelation of the state doesn't need to be exact match anymore.